+++ frame_loop.odin
@@ -9,7 +9,7 @@ quit_app := false
 
 frame_loop :: proc(window: ^SDL.Window, gs: ^Game_State, rs: ^RenderState, ss: ^ShaderState, ps: ^Physics_State) {
     TARGET_FRAME_RATE :: 60.0
-    FIXED_DELTA_TIME :: 1.0 / TARGET_FRAME_RATE
+    FIXED_DELTA_TIME :: f32(1.0 / TARGET_FRAME_RATE)
     clocks_per_second := SDL.GetPerformanceFrequency()
     target_frame_clocks := clocks_per_second / TARGET_FRAME_RATE
     max_deviation: u64 = clocks_per_second / 4000
